+++ packages/ui/src/icons/index.ts
@@ -51,3 +51,4 @@ export * from './ShareIcon';
 export * from './NoCollectiblesIcon';
 export * from './ExternalLinkIcon';
 export * from './NoImageIcon';
+export * from './SecurityIcon';